+++ CMakeLists.txt
@@ -199,15 +199,17 @@ target_sources(${PROGRAM_NAME} PRIVATE
     "${SCHETOOL_SRC}/${HANDLERS_DIR}/window.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/autosave_popup_gui.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/edit_history_gui.cpp"
-    "${SCHETOOL_SRC}/${INTERFACE_DIR}/element_editor_subgui.cpp"
-    "${SCHETOOL_SRC}/${INTERFACE_DIR}/filter_editor_subgui.cpp"
+    "${SCHETOOL_SRC}/${INTERFACE_DIR}/schedule/element_editor_subgui.cpp"
+    "${SCHETOOL_SRC}/${INTERFACE_DIR}/schedule/filter_editor_subgui.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/gui.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/gui_templates.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/interface.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/interface_style.cpp"
-    "${SCHETOOL_SRC}/${INTERFACE_DIR}/main_menu_bar_gui.cpp"
-    "${SCHETOOL_SRC}/${INTERFACE_DIR}/schedule_gui.cpp"
+    "${SCHETOOL_SRC}/${INTERFACE_DIR}/main_menu_bar/delete_modal_subgui.cpp"
+    "${SCHETOOL_SRC}/${INTERFACE_DIR}/main_menu_bar/main_menu_bar_gui.cpp"
+    "${SCHETOOL_SRC}/${INTERFACE_DIR}/schedule/schedule_gui.cpp"
     "${SCHETOOL_SRC}/${INTERFACE_DIR}/start_page_gui.cpp"
+    "${SCHETOOL_SRC}/${INTERFACE_DIR}/text_input_modal_subgui.cpp"
     "${SCHETOOL_SRC}/${IO_DIR}/io_handler.cpp"
     "${SCHETOOL_SRC}/${IO_DIR}/preferences_data_converter.cpp"
     "${SCHETOOL_SRC}/${IO_DIR}/preferences_io.cpp"
