+++ include/interface/main_menu_bar/delete_modal_subgui.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "text_input_modal_subgui.h"
+
+class DeleteModalSubGui : public Gui {
+    private:
+        std::string m_deleteConfirmationScheduleName = "";
+
+    public:
+        DeleteModalSubGui(const char* ID) : Gui(ID) {
+        }
+
+        Event<std::string> deleteScheduleEvent;
+
+        void draw(Window& window, Input& input, GuiTextures& guiTextures) override;
+        void setAffectedScheduleName(const std::string& name);
+};
