+++ include/interface/start_page_gui.h
@@ -2,20 +2,10 @@
 
 #include <vector>
 #include "event.h"
+#include "event_pipe.h"
 #include "gui.h"
 #include "window.h"
 #include "input.h"
-#include "main_menu_bar_gui.h"
-
-class StartPageNewNameModalSubGui : public TextInputModalSubGui {
-    public:
-        StartPageNewNameModalSubGui(const char* ID) : TextInputModalSubGui(ID, "Enter name   ", "Create schedule") {
-        }
-
-        Event<std::string> createNewScheduleEvent;
-
-        void invokeEvent(const std::string& text) override;
-};
 
 class StartPageGui : public Gui {
     private:
@@ -26,6 +16,8 @@ class StartPageGui : public Gui {
         StartPageGui(const char* ID);
 
         Event<std::string> openScheduleFileEvent;
+        // Event pipes
+        EventPipe<std::string> createNewScheduleEventPipe;
 
         void draw(Window& window, Input& input, GuiTextures& guiTextures) override;
         void passFileNames(const std::vector<std::string>& fileNames);
