+++ source/interface/main_menu_bar/main_menu_bar_gui.cpp
@@ -1,109 +1,21 @@
-#include "main_menu_bar_gui.h"
+#include "main_menu_bar/main_menu_bar_gui.h"
+#include "main_menu_bar/delete_modal_subgui.h"
+#include "text_input_modal_subgui.h"
 #include "gui_templates.h"
 #include "gui_constants.h"
 #include "util.h"
 
-TextInputModalSubGui::TextInputModalSubGui(
-    const char* ID, const char* popupName, const char* acceptButtonText, size_t textMaxLength, bool showCloseButton)
-    : Gui(ID) {
-    m_popupName = popupName;
-    m_acceptButtonText = acceptButtonText;
-    m_textMaxLength = textMaxLength;
-    m_showCloseButton = showCloseButton;
-}
-
-void TextInputModalSubGui::draw(Window& window, Input& input, GuiTextures& guiTextures) {
-    float nameInputWidth = ImGui::CalcTextSize(std::string(m_textMaxLength + 2, 'a').c_str()).x;
-
-    m_visible = true;
-
-    ImGuiStyle& style = ImGui::GetStyle();
-    ImGui::SetNextWindowSize(ImVec2(nameInputWidth + style.FramePadding.x * 2.0f, 100));
-    // ImGui::SetNextWindowPos(ImVec2(0.0f, 0.0f), 0, ImVec2(0.5f, 0.5f));
-    if (ImGui::BeginPopupModal(
-            m_popupName, m_showCloseButton ? &m_visible : NULL, ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove))
-    {
-        m_textBuffer.reserve(m_textMaxLength);
-        char* buf = m_textBuffer.data();
-        ImGui::PushItemWidth(nameInputWidth - style.FramePadding.x * 2.0f);
-        ImGui::InputText(std::string("##TextInput").append(m_popupName).c_str(),
-                         buf,
-                         m_textMaxLength,
-                         ImGuiInputTextFlags_CallbackCharFilter,
-                         gui_callbacks::filterAlphanumerics);
-        ImGui::PopItemWidth();
-        if (std::string(buf).empty() == false) {
-            float size = 128.0f + style.FramePadding.x * 2.0f;
-            float avail = ImGui::GetContentRegionAvail().x;
-            float off = (avail - size) * 0.5f;
-            if (off > 0.0f) {
-                ImGui::SetCursorPosX(ImGui::GetCursorPosX() + off);
-            }
-            if (ImGui::Button(m_acceptButtonText.c_str(), ImVec2(128, 0))) {
-                invokeEvent(std::string(buf));
-            }
-        }
-        ImGui::EndPopup();
-    }
-}
-
-// Open the popup and empty the text buffer
-void TextInputModalSubGui::open() {
-    m_textBuffer = "";
-    ImGui::OpenPopup(m_popupName);
-}
-
-void TextInputModalSubGui::invokeEvent(const std::string& text) {
-}
-
-void NewNameModalSubGui::invokeEvent(const std::string& text) {
-    createNewScheduleEvent.invoke(text);
-}
-
-void RenameModalSubGui::invokeEvent(const std::string& text) {
-    renameScheduleEvent.invoke(text);
-}
-
-void DeleteModalSubGui::draw(Window& window, Input& input, GuiTextures& guiTextures) {
-    ImGui::SetNextWindowSize(ImVec2(386.0f, 100.0f));
-    if (ImGui::BeginPopupModal("Confirm Schedule deletion", NULL, ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove)) {
-        ImGuiStyle& style = ImGui::GetStyle();
-        ImGui::Text("%s %s?", "Delete", m_deleteConfirmationScheduleName.c_str());
-        ImGui::NewLine();
-        float size = 120.0f + style.FramePadding.x * 2.0f;
-        float avail = ImGui::GetContentRegionAvail().x;
-        float off = (avail - size) * 0.23f;
-        if (off > 0.0f) {
-            ImGui::SetCursorPosX(ImGui::GetCursorPosX() + off);
-        }
-        if (ImGui::Button("Delete##DeleteSchedule", ImVec2(120, 0))) {
-            deleteScheduleEvent.invoke(std::string(m_deleteConfirmationScheduleName));
-
-            ImGui::CloseCurrentPopup();
-        }
-        ImGui::SameLine();
-        // janky offset but it works so i don't much care
-        off = (avail - size) * 0.04f;
-        if (off > 0.0f) {
-            ImGui::SetCursorPosX(ImGui::GetCursorPosX() + off);
-        }
-        if (ImGui::Button("Cancel##CancelScheduleDeletion", ImVec2(120, 0))) {
-            ImGui::CloseCurrentPopup();
-        }
-        ImGui::EndPopup();
-    }
-}
-
-void DeleteModalSubGui::setAffectedScheduleName(const std::string& name) {
-    m_deleteConfirmationScheduleName = name;
-}
-
 MainMenuBarGui::MainMenuBarGui(const char* ID, std::shared_ptr<const InterfaceStyleHandler> styleHandler) : Gui(ID) {
     m_styleHandler = styleHandler;
-    // add subguis
-    addSubGui(new NewNameModalSubGui("NewNameModalSubGui"));
-    addSubGui(new RenameModalSubGui("RenameModalSubGui"));
+    // Add subguis
     addSubGui(new DeleteModalSubGui("DeleteModalSubGui"));
+    addSubGui(new TextInputModalSubGui("NewNameModalSubGui", "Enter name", "Create schedule"));
+    addSubGui(new TextInputModalSubGui("RenameModalSubGui", "Enter new name"));
+
+    // Link event pipes
+    createNewScheduleEventPipe.addEvent(getSubGui<TextInputModalSubGui>("NewNameModalSubGui")->acceptButtonPressedEvent);
+    deleteScheduleEventPipe.addEvent(getSubGui<DeleteModalSubGui>("DeleteModalSubGui")->deleteScheduleEvent);
+    renameScheduleEventPipe.addEvent(getSubGui<TextInputModalSubGui>("RenameModalSubGui")->acceptButtonPressedEvent);
 }
 
 // Helper function that gets the input shortcuts for an INPUT_EVENT and turns them into a string in the format:
@@ -204,7 +116,7 @@ void MainMenuBarGui::draw(Window& window, Input& input, GuiTextures& guiTextures
         newSchedule();
     }
 
-    if (auto newNameModalSubGui = getSubGui<NewNameModalSubGui>("NewNameModalSubGui")) {
+    if (auto newNameModalSubGui = getSubGui<TextInputModalSubGui>("NewNameModalSubGui")) {
         newNameModalSubGui->draw(window, input, guiTextures);
         if (m_openNewNameModal) {
             newNameModalSubGui->open();
@@ -212,7 +124,7 @@ void MainMenuBarGui::draw(Window& window, Input& input, GuiTextures& guiTextures
         }
     }
 
-    if (auto renameModalSubGui = getSubGui<RenameModalSubGui>("RenameModalSubGui")) {
+    if (auto renameModalSubGui = getSubGui<TextInputModalSubGui>("RenameModalSubGui")) {
         renameModalSubGui->draw(window, input, guiTextures);
     }
